requirements.txt
# FastAPI and ASGI server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Database
sqlalchemy>=2.0.36
aioodbc>=0.5.0
pyodbc>=5.1.0

# Security and Authentication
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.12

# Machine Learning
numpy>=1.26.0
scikit-learn>=1.5.0

# Environment management
python-dotenv>=1.0.1

